requirements.txt
alembic==1.8.1
anyio==3.6.1
click==8.1.3
fastapi==0.79.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
Mako==1.2.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pydantic==1.9.1
python-dotenv==0.20.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
